remove "items". poorly.
