+++ app/routes/room.js
@@ -2,9 +2,6 @@ import Ember from 'ember';
 
 export default Ember.Route.extend({
   inventory: Ember.inject.service('inventory'),
-  model: function(params){
-    return this.store.find('room', params.room_id);
-  },
   afterModel: function(model, transition){
     var inventory = this.get('inventory');
     if(model.get('locked') && !inventory.hasKey(model.get('id'))){
@@ -15,6 +12,14 @@ export default Ember.Route.extend({
   actions: {
     didTransition: function(){
       this.get('inventory').set('message', '');
+    },
+    takeKey: function(key){
+      this.get("inventory.keys").pushObject(key);
+      this.currentModel.set("key", null);
+    },
+    takeThing: function(thing){
+      this.get("inventory.things").pushObject(thing);
+      this.currentModel.get("things").removeObject(thing);
     }
   }
 });
