--- app/routes/room/index.js
@@ -1,7 +1,0 @@
-import Ember from 'ember';
-
-export default Ember.Route.extend({
-  model: function(){
-    return this.modelFor('room');
-  }
-});
