--- app/templates/room/index.hbs
@@ -1,8 +1,0 @@
-<p class="room-description">you are {{model.description}}.
-  {{#if model.items}}
-    you see
-    {{#each model.items as |item|}}
-      {{link-to item.name 'room.item' model item.id class="item"}}.
-    {{/each}}
-  {{/if}}
-</p>
