--- app/templates/room/item.hbs
@@ -1,19 +1,0 @@
-<p class="item-description">
-  {{model.item.name}}...{{model.item.description}}
-</p>
-<p class="item-contents">
-  {{#if model.item.key}}
-    There is
-    <a class="key" href="" {{action "takeKey" model.item.key}}>
-      {{model.item.key.description}}
-    </a>
-    inside.
-  {{/if}}
-  {{#each model.item.things as |thing|}}
-    You see
-    <a class="thing" href="" {{action "takeThing" thing}}>
-      {{thing.name}}
-    </a>
-  {{/each}}
-</p>
-{{link-to 'back' 'room' model.room class="back"}}
