--- tests/unit/models/item-test.js
@@ -1,12 +1,0 @@
-import {
-  moduleForModel,
-  test
-} from 'ember-qunit';
-
-moduleForModel('item', {integration: true});
-
-test('it exists', function(assert) {
-  var model = this.subject();
-  // var store = this.store();
-  assert.ok(!!model);
-});
